Replace "xcb" with "as-raw-xcb-connection"
There is a new crate "as-raw-xcb-connection" which provides a trait that
can be used to get a raw pointer to a libxcb xcb_connection_t. This
trait allows this crate to work with arbitrary versions of the xcb crate
and other crates wrapping libxcb (as long as they implement this trait).

Signed-off-by: Uli Schlachter <psychon@znc.in>